# mini_blog_server/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/mini/api/article/get/category": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ArticleAPI"
                ],
                "summary": "获取文章分类",
                "responses": {
                    "200": {
                        "description": "{\"code\":200,\"msg\":\"success\",\"data\":{}}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/mini/api/article/get/detail": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ArticleAPI"
                ],
                "summary": "获取文章详情",
                "parameters": [
                    {
                        "description": "文章href链接",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.ArticleDetail"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"code\":200,\"msg\":\"success\",\"data\":{}}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/mini/api/article/get/list": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ArticleAPI"
                ],
                "summary": "获取文章列表",
                "parameters": [
                    {
                        "description": "分页",
                        "name": "pageInfo",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.PageInfo"
                        }
                    },
                    {
                        "description": "文章分类",
                        "name": "cateInfo",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.ArticleCategory"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"code\":200,\"msg\":\"success\",\"data\":{}}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/mini/api/article/get/recommend": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ArticleAPI"
                ],
                "summary": "获取推荐文章",
                "responses": {
                    "200": {
                        "description": "{\"code\":200,\"msg\":\"success\",\"data\":{}}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "request.ArticleCategory": {
            "type": "object",
            "properties": {
                "cateId": {
                    "type": "integer"
                }
            }
        },
        "request.ArticleDetail": {
            "type": "object",
            "properties": {
                "href": {
                    "type": "string"
                }
            }
        },
        "request.PageInfo": {
            "type": "object",
            "properties": {
                "page": {
                    "type": "integer"
                },
                "pageSize": {
                    "type": "integer"
                }
            }
        }
    }
}